+++ src/EloquentDataTable.php
@@ -1,9 +1,0 @@
-<?php
-
-namespace ElfSundae\Laravel\DataTables;
-
-use Yajra\DataTables\EloquentDataTable as BaseEloquentDataTable;
-
-class EloquentDataTable extends BaseEloquentDataTable
-{
-}